requirements.txt
Django==4.1.6
django-crispy-forms==1.14.0
djangorestframework==3.14.0
